issue-2 - Desenvolver um CRUD de exemplo
Implementado CRUD no fluxo de exemplo para demonstrar a base de como realizar todas as operações de banco de dados

+++ castor-core/src/main/java/br/com/restaurante/castor/repositories/ExamplePessoaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ExamplePessoaRepository extends JpaRepository<Pessoa, Integer>
+public interface ExamplePessoaRepository extends JpaRepository<Pessoa, Long>
 {
     Optional<Pessoa> findByName(String name);
 }
